update: tx history query, sender OR recipient match
