+++ src/utils/index.ts
@@ -70,16 +70,9 @@ export function toBytes32(addr: string): string {
   if (addr.length === 42) {
     // trim 12 bytes from beginning plus '0x'
     const short = addr.slice(2)
-    console.log('short', short)
     return `0x000000000000000000000000${short}`
   }
   throw new Error('invalid address length, cannot convert to 20 bytes')
-}
-
-export function toBytes32(addr: string): string {
-  // trim 12 bytes from beginning plus '0x'
-  const short = addr.slice(2)
-  return `0x000000000000000000000000${short}`
 }
 
 /**
